+++ packages/user/src/supertokens/types.ts
@@ -1,6 +1,6 @@
 import ThirdPartyEmailPassword from "supertokens-node/recipe/thirdpartyemailpassword";
 
-import type { EmailVerificationRecipe } from "./emailVerificationRecipeTypes";
+import type { EmailVerificationRecipe } from "./emailVerificationRecipeConfigTypes";
 import type { FastifyInstance } from "fastify";
 import type { EmailDeliveryInterface } from "supertokens-node/lib/build/ingredients/emaildelivery/types";
 import type { TypeEmailPasswordPasswordResetEmailDeliveryInput } from "supertokens-node/lib/build/recipe/emailpassword/types";
